% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthforum-package.R
\docType{package}
\name{healthforum-package}
\alias{healthforum}
\alias{healthforum-package}
\title{healthforum: Scrape Patient Forum Data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Scrape data from Patient Forum <https://patient.info/forums> by entering urls. It will return a data frame containing text, user names, like counts, reply counts, etc.
}
\author{
\strong{Maintainer}: Lingshu Hu \email{lingshu.hu@hotmail.com} (0000-0003-0304-882X)

Other contributors:
\itemize{
  \item Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694) [contributor]
}

}
\keyword{internal}
