% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medical_words}
\alias{medical_words}
\title{The English medical glossary dictionary}
\format{A data frame with 94959 rows and 2 variables:
\describe{
  \item{word}{medical words}
  \item{value}{weight of the medical words}
  ...
}}
\source{
\url{https://github.com/glutanimate/wordlist-medicalterms-en}
}
\usage{
medical_words
}
\description{
A dataset containing two prominent medical dictionary projects
OpenMedSpel by R. Robinson of e-MedTools (Version 2.0.0, released 2014-01-21)
<http://www.e-medtools.com/openmedspel.html> and
MTH-Med-Spel-Chek by Rajasekharan N. of MT-Herald (released 2014-04-02)
<http://mtherald.com/free-medical-spell-checker-for-microsoft-word-custom-dictionary/>.
}
\keyword{datasets}
