% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_data.R
\name{scrape_groups_by_initial_letter}
\alias{scrape_groups_by_initial_letter}
\title{Scrape groups by initial letter}
\usage{
scrape_groups_by_initial_letter(index, post_number_per_group = NULL, ...)
}
\arguments{
\item{index}{The initial letter of groups. Can be one letter or a vector of letters.}

\item{post_number_per_group}{The number of random posts to scrape per group. Default is NULL, which means scrape the total number of posts in each group}

\item{...}{optional arguments to FUN.}
}
\value{
A data frame
}
\description{
Get posts and all the replies to the posts from groups by entering the initial letter of group names
}
\examples{

\donttest{
## Get the posts data of groups whose names starting with the letter "x".
scrape_groups_by_initial_letter(index = "x", post_number_per_group = 1)
}

}
