% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_one_group.R
\name{scrape_one_group}
\alias{scrape_one_group}
\title{Scrape one group}
\usage{
scrape_one_group(group_url, random_post_number = NULL)
}
\arguments{
\item{group_url}{URL to the page to scrape}

\item{random_post_number}{Random number of posts to scrape}
}
\value{
A data frame
}
\description{
Get all the posts and all the replies to the posts from one group by entering its url
}
