% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_data.R
\name{scrape_one_post}
\alias{scrape_one_post}
\title{Scrape one initial post}
\usage{
scrape_one_post(url, From = 1L, To = length(urls))
}
\arguments{
\item{url}{URL to the post to scrape}

\item{From}{The starting page number. Default is the first page}

\item{To}{The ending page number. Default is the last page}
}
\value{
A data frame
}
\description{
Get the data from one initial post by entering its url
}
\examples{
## get two pages of data from the post titled "Can Gastritis be cured?"
scrape_one_post(url = "https://patient.info/forums/discuss/can-gastritis-be-cured--613999",
                From = 1, To = 2)

}
