% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_data.R
\name{scrape_user_posts}
\alias{scrape_user_posts}
\title{Scrape a user's posts}
\usage{
scrape_user_posts(user_profile_url, type = c("both", "replies",
  "topic_post"))
}
\arguments{
\item{user_profile_url}{The URL of a user's profile page. Multiple users use c(url1, url2, ...)}

\item{type}{Choose a type of posts to be scraped. It includes "replies", i.e., a user's replies to others' posts, or "topic_post", i.e., a user's initial posts. The default is to get "both"}
}
\value{
A data frame
}
\description{
Get all posts/replies one user has sent by his/her user name
}
\examples{
user_profile_url <- "https://patient.info/forums/profiles/utgh4k33-1264038"
scrape_user_posts(user_profile_url = user_profile_url, type = "both")

}
